Refactor: refactored the code base for a final push to dev
