+++ app/(searching-consultant)/search.js
@@ -4,9 +4,9 @@ import { StyleSheet, ScrollView, View } from 'react-native';
 import { theme } from '../../styles/theme';
 import { useSpecialities } from '../../context/FilterConsultantsContext';
 import SearchBar from '../../components/common/SearchBar';
-import FilterTabs from '../../components/searching-consultant/FilterTabs';
+import FilterTabs from '../../components/searching-consultant/tabs/FilterTabs';
 import ConsultantCard from '../../components/searching-consultant/consultant-card/ConsultantCard';
-import useConsultants from '../../services/hooks/useConsultants';
+import useConsultants from '../../hooks/useConsultants';
 import placeHolder from '../../locales/en/SearchBar.json';
 
 const { consultingSearching } = placeHolder;
