+++ components/searching-consultant/common/ModalHeader.js
@@ -1,18 +1,20 @@
 import React from 'react';
 import { View, Text, StyleSheet, Pressable } from 'react-native';
-
+import tabs from '../../../locales/en/Tabs.json';
+//This component is in every modal. It is header with a description of what is for and a clear all button.
+//Onpress is usually the Clear all button, and filter is the label of the Modal.
 const ModalHeader = ({ onPress, filter }) => {
 	return (
 		<View style={styles.header}>
 			<Text style={styles.headerText}>{filter}</Text>
 			<Pressable onPress={onPress}>
-				<Text style={styles.clearAllText}>CLEAR ALL</Text>
+				<Text style={styles.clearAllText}>{tabs.ClearAll}</Text>
 			</Pressable>
 		</View>
 	);
 };
 
-import { theme } from '../../../../styles/theme';
+import { theme } from '../../../styles/theme';
 
 const styles = StyleSheet.create({
 	header: {
@@ -21,7 +23,6 @@ const styles = StyleSheet.create({
 		justifyContent: 'space-between',
 	},
 	headerText: {
-		color: '#000',
 		letterSpacing: 0.1,
 		fontFamily: 'Roboto-Bold',
 		fontSize: theme.typography.large.fontSize,
