+++ components/searching-consultant/filters/filters-tabs/PriceRangeModal.js
@@ -2,22 +2,31 @@ import React from 'react';
 import { Text, StyleSheet, View } from 'react-native';
 import Slider from '@react-native-community/slider';
 
-import { useSpecialities } from '../../../context/FilterConsultantsContext';
-import ModalContent from './ModalContent';
+import { useSpecialities } from '../../../../context/FilterConsultantsContext';
+import ModalContent from '../ModalContent';
+import tabs from '../../../../locales/en/Tabs.json';
 
 const PriceRangeModalContent = ({ closeModal }) => {
 	const { price, setPrice } = useSpecialities();
 
+	const handleSliderValueChange = (value) => {
+		setPrice(value);
+	};
+
 	const scrollViewContent = (
 		<View style={styles.sliderContainer}>
-			<Text style={styles.sliderLabel}>$ {price}</Text>
+			<Text style={styles.sliderLabel}>
+				{tabs.PriceRange.text}
+				{price}
+			</Text>
 			<Slider
 				style={styles.slider}
 				minimumValue={0}
 				maximumValue={30}
 				step={1}
 				value={price}
-				onValueChange={(value) => setPrice(value)}
+				onValueChange={handleSliderValueChange}
+				onSlidingComplete={handleSliderValueChange}
 			/>
 		</View>
 	);
@@ -26,13 +35,13 @@ const PriceRangeModalContent = ({ closeModal }) => {
 		<ModalContent
 			closeModal={closeModal}
 			headerAction={() => setPrice(0)}
-			headerFilter={'Price'}
+			headerFilter={tabs.PriceRange.label}
 			scrollViewContent={scrollViewContent}
 		/>
 	);
 };
 
-import { theme } from '../../../styles/theme';
+import { theme } from '../../../../styles/theme';
 
 const styles = StyleSheet.create({
 	sliderContainer: {
