+++ components/searching-consultant/filters/filters-tabs/RegionModal.js
@@ -1,0 +1,91 @@
+import React from 'react';
+import { Text, StyleSheet, View } from 'react-native';
+
+import { useSpecialities } from '../../../../context/FilterConsultantsContext';
+import philippineRegions from '../../../../defaultdata/Philippines-Region.json';
+import CheckBox from '../../../../assets/svg/CheckIcon';
+import ModalContent from '../ModalContent';
+import tabs from '../../../../locales/en/Tabs.json';
+
+const RegionItem = ({ region, onPress, isSelected }) => (
+	<View style={styles.regionContainer}>
+		<Text
+			onPress={onPress}
+			style={[styles.regionText, isSelected && styles.selectedRegionText]}
+		>
+			{region}
+		</Text>
+		{isSelected && (
+			<CheckBox
+				width={theme.spacing.xlarge}
+				height={theme.spacing.xlarge}
+				color={theme.colors.text.dark}
+				isChecked={true}
+			/>
+		)}
+	</View>
+);
+
+const RegionList = ({ regions, onRegionPress, selectedRegions }) => (
+	<>
+		{regions.map((region, index) => (
+			<RegionItem
+				key={index}
+				region={region}
+				onPress={() => onRegionPress(region)}
+				isSelected={selectedRegions.includes(region)}
+			/>
+		))}
+	</>
+);
+
+const RegionModalContent = ({ closeModal }) => {
+	const { addRegion, removeRegion, selectedRegions, clearRegions } =
+		useSpecialities();
+
+	const toggleRegion = (region) => {
+		if (selectedRegions.includes(region)) {
+			removeRegion(region);
+		} else {
+			addRegion(region);
+		}
+	};
+
+	return (
+		<ModalContent
+			closeModal={closeModal}
+			headerAction={clearRegions}
+			headerFilter={tabs.Region}
+			scrollViewContent={
+				<RegionList
+					regions={philippineRegions}
+					onRegionPress={toggleRegion}
+					selectedRegions={selectedRegions}
+				/>
+			}
+		/>
+	);
+};
+
+import { theme } from '../../../../styles/theme';
+
+const styles = StyleSheet.create({
+	regionContainer: {
+		flexDirection: 'row',
+		alignItems: 'center',
+		flex: 1,
+	},
+	regionText: {
+		color: theme.colors.text.dark,
+		letterSpacing: 0.1,
+		flexGrow: 1,
+		fontFamily: 'Roboto-Regular',
+		fontSize: theme.typography.mediumBody.fontSize,
+		paddingVertical: theme.spacing.medium,
+	},
+	selectedRegionText: {
+		color: theme.colors.secondary.dark,
+	},
+});
+
+export default RegionModalContent;
