+++ defaultdata/Filtering-Tabs.json
@@ -1,5 +1,5 @@
 [
-    "Speciality",
+    "Specialty",
     "Experience",
     "Region",
     "Price Range"
